오류수정

+++ WinApiLecture/Client/CScene_Start.cpp
@@ -102,10 +102,10 @@ void CScene_Start::render(HDC _dc)
 	Vec2 vRenderPos=CCamera::GetInst()->GetRenderPos(m_vForcePos);
 
 	Ellipse(_dc
-		, vRenderPos.x - m_fCurRadius
-		, vRenderPos.y - m_fCurRadius
-		, vRenderPos.x + m_fCurRadius
-		, vRenderPos.y + m_fCurRadius);
+		, (int)(vRenderPos.x - m_fCurRadius)
+		, (int)(vRenderPos.y - m_fCurRadius)
+		, (int)(vRenderPos.x + m_fCurRadius)
+		, (int)(vRenderPos.y + m_fCurRadius));
 
 }
